+++ src/components/calendar/index.js
@@ -221,7 +221,7 @@ Calendar.propTypes = {
     disabledDates: PropTypes.array,
     minDate: PropTypes.any,
     maxDate: PropTypes.any,
-    selectBorderType: PropTypes.oneOfType([undefined,'','default','tile','rounded','smooth']),
+    selectBorderType: PropTypes.oneOf([undefined,'','default','tile','rounded','smooth']),
     onDate: PropTypes.func,
     hideWeekend: PropTypes.bool,
     hideCurrentDay: PropTypes.bool,